% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_S3.R
\name{angle}
\alias{angle}
\title{Constructor for an S3 object of class "angle"}
\usage{
angle(ang, rad_flag = TRUE)
}
\arguments{
\item{ang}{A real number, in degrees or radians depending on rad_flag.}

\item{rad_flag}{A logical flag. If FALSE, the value is meant to be in radians.}
}
\value{
An object of class "angle" whose numerical value is always in degrees.
}
\description{
Constructor for an S3 object of class "angle"
}
\examples{
# Create an angle of 60 degrees
ang1 <- angle(60)
class(ang1)
}
