% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections.R
\name{aveIvsRes}
\alias{aveIvsRes}
\title{Mean and standard deviation in resolution shells.}
\usage{
aveIvsRes(nbin, resos, II = NULL, m = max(resos), M = min(resos))
}
\arguments{
\item{nbin}{A positive integer. The number of resolution shells.}

\item{II}{A vector of real quantities. This is the key quantity whose averages
and standard deviations are calculated. If \code{II} is set to \code{NULL}, resolutions
averages and standard deviations will be the calculated quantities.}
}
\value{
A named list of length 4. \code{counts} is a vector of integers, the number of
 reflections in each resolution shell. \code{mids} is the representative inverse resolution
 for each resolution shell; the value is decided by the function \code{hist}. \code{ave}
 is the average value in each resolution shell and \code{sd} is the corresponding standard
 deviation.
}
\description{
Calculates averages and standard deviations of the input vector quantity for all
reflections, corresponding to shells of resolution.
}
\details{
Binning is done with inverse resolutions in order to have lower resolutions
 correspond to small numbers and high resolutions to large numbers. The output
 \code{mids}, \code{ave} and \code{sd} correspond to inverse resolutions.
}
\examples{
datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"insulin_merged.mtz")
lmtz <- readMTZ(filename)
hkl <- lmtz$reflections[,1:3]
II <- lmtz$reflections[,4]
cpars <- lmtz$header$CELL
resos <- hkl_to_reso(hkl[,1],hkl[,2],hkl[,3],
                     cpars[1],cpars[2],cpars[3],
                     cpars[4],cpars[5],cpars[6])
ltmp <- aveIvsRes(20,resos,II)
plot(ltmp$mids,ltmp$ave,type="b",pch=16)
}
