% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_angle_validity}
\alias{check_angle_validity}
\title{Validity and compatibility of a cry object of class 'angle'}
\usage{
check_angle_validity(x)
}
\arguments{
\item{x}{Object of class angle.}
}
\value{
ans A logical value. TRUE means that the input is a valid object of class'angle'.
}
\description{
An object of class 'angle' is a numeric with logical attribure "rad_flag".
}
\examples{
# Create an object of class angle
x <- angle(80)

# Check its validity
check_angle_validity(x)

# Modify the 'rad_flag' attribute
attr(x,"rad_flag") <- 12.5

# Check its validity
check_angle_validity(x)

}
