% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_bravais_validity}
\alias{check_bravais_validity}
\title{Validity and compatibility of a cry object of class 'bravais'}
\usage{
check_bravais_validity(x)
}
\arguments{
\item{x}{Object of class 'bravais'.}
}
\value{
ans A logical value. TRUE means that the input is a valid object of class'bravais'.
}
\description{
An object of class 'bravais' is a named list of length 4. The first slot, "bt", is
the universally-used two-letter symbol. The second, third and fourth slots are, respectively,
"cr_fam" (the corresponding crystal family), "cr_sys" (the corresponding crystal system) and
"lt_sys" (the corresponding lattice system).
}
\examples{
# Create an object of class 'bravais'
x <- bravais("mP")

# Check its validity
check_bravais_validity(x)

}
