% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_rec_unit_cell_validity}
\alias{check_rec_unit_cell_validity}
\title{Validity and compatibility of a cry object of class 'rec_unit_cell'}
\usage{
check_rec_unit_cell_validity(x)
}
\arguments{
\item{x}{Object of class 'rec_unit_cell'.}
}
\value{
ans A logical value. TRUE means that the input is a valid object of class'rec_unit_cell'.
}
\description{
An object of class 'rec_unit_cell' is a named list of length 6. The first three fields are
numeric, the last three of class 'angle'.
}
\examples{
# Create an object of class 'rec_unit_cell'
x <- create_rec_unit_cell()

# Check its validity
check_rec_unit_cell_validity(x)

# Now change two fields
x$ar <- "ar"
x$alphar <- 123

# Check validity again
check_rec_unit_cell_validity(x)

}
