% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_S3.R
\name{CIF}
\alias{CIF}
\title{Constructor for an S3 object of class "CIF".}
\usage{
CIF(filename = NULL)
}
\arguments{
\item{filename}{A string character. The name of the CIF file containing the data.}
}
\value{
An object of class "CIF". It's a named list whose number and name of fields
 depends on the specific information content of the file.
}
\description{
This represents a CIF (Crystallographic Information Framework) data file.
}
\details{
Crystallographic data can be collected in CIF files in the most complete and
structured way. An object of this class is created when the input is the name
of a valid CIF file. The output is a named list with number of fields and names
dependent on the CIF information included in the specific file.
}
\examples{
# Read a CIF file

}
