% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{create_rec_unit_cell.CIF}
\alias{create_rec_unit_cell.CIF}
\title{Reciprocal unit cell starting from a CIF object}
\usage{
\method{create_rec_unit_cell}{CIF}(ar, ...)
}
\arguments{
\item{ar}{An object of class "CIF".}

\item{...}{Additional arguments passed to the create_rec_unit_cell methods}
}
\value{
An object of class "rec_unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
Method to create an object of class "rec_unit_cell" starting from an object of
class "CIF".
}
\examples{
# Create a CIF object from a cif file
#cobj <- CIF("directory/filename.cif")

# Create rec_unit_cell object
#ruc <- create_rec_unit_cell(cobj)
#print(ruc)

}
