% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cry_generics_S3.R
\name{create_rec_unit_cell}
\alias{create_rec_unit_cell}
\title{S3 generic to create rec_unit_cell objects}
\usage{
create_rec_unit_cell(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The rec_unit_cell object can be created starting from specific objects, files, etc.
}
\examples{
# Create a rec_unit_cell in default (no arguments)
ruc <- create_rec_unit_cell()
print(ruc)
}
