% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{create_rec_unit_cell.unit_cell}
\alias{create_rec_unit_cell.unit_cell}
\title{Reciprocal unit cell starting from a unit cell}
\usage{
\method{create_rec_unit_cell}{unit_cell}(uc)
}
\arguments{
\item{uc}{An object of class "unit_cell".}
}
\value{
An object of class "rec_unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
Method to create an object of class "rec_unit_cell" starting from an object of
class "unit_cell".
}
\examples{
# Create a "rec_unit_cell" object starting from a cubic cell object
uc <- unit_cell()
print(uc)
ruc <- create_rec_unit_cell(uc)
print(ruc)

}
