% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{create_unit_cell.bravais}
\alias{create_unit_cell.bravais}
\title{Method to create a unit_cell object starting from a Bravais symbol}
\usage{
\method{create_unit_cell}{bravais}(bt, cpar = NULL)
}
\arguments{
\item{bt}{An object of class "bravais".}
}
\value{
An object of class "unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
The Bravais symbols indicate the 14 possible Bravais lattices. A few
examples are "aP", "oF", etc. The cell parameters assigned are assigned
randomly, but are compatible with the Bravais lattice.
}
\examples{
# Create a "unit_cell" object from a monoclinic primitive Bravais lattice
# Cell parameters generated automatically.
bt <- bravais("mP")
uc <- create_unit_cell(bt)
print(uc)

# Create a "unit_cell" object from a monoclinic primitive Bravais lattice
# Cell parameters fixed by user.
bt <- bravais("mP")
uc <- create_unit_cell(bt,c(10,30,20,90,105,90))
print(uc)
}
