% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{create_unit_cell.CIF}
\alias{create_unit_cell.CIF}
\title{Unit cell starting from a CIF object}
\usage{
\method{create_unit_cell}{CIF}(a, ...)
}
\arguments{
\item{a}{An object of class "CIF".}

\item{...}{Additional arguments passed to the create_unit_cell methods}
}
\value{
An object of class "unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
Method to create an object of class "unit_cell" starting from an object of
class "CIF".
}
\examples{
# Create an object of class 'CIF' from a file
#cobj <- CIF("directory.filename.cif")

# Unit_cell object from CIF object
#uc <- create_unit_cell(cobj)
#print(uc)

}
