% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{create_unit_cell.rec_unit_cell}
\alias{create_unit_cell.rec_unit_cell}
\title{Unit cell starting from a reciprocal unit cell}
\usage{
\method{create_unit_cell}{rec_unit_cell}(a, ...)
}
\arguments{
\item{a}{An object of class "rec_unit_cell".}

\item{...}{Additional arguments passed to the create_unit_cell methods}
}
\value{
An object of class "unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
Method to create an object of class "unit_cell" starting from an object of
class "rec_unit_cell".
}
\examples{
# Create a "unit_cell" object starting from a reciprocal cubic cell object
ruc <- rec_unit_cell()
print(ruc)
uc <- create_unit_cell(ruc)
print(uc)

}
