% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clattice.R
\name{crystal_family}
\alias{crystal_family}
\title{Crystal family corresponding to given space group.}
\usage{
crystal_family(gn)
}
\arguments{
\item{gn}{A natural integer (1,2,3,...). the space group number.}
}
\value{
A character string, the name of the crystal family associated
        with the given space group. If the input integer does not correspond
        any existing space group, the function returns NULL and throws a
        warning.
}
\description{
Crystal family corresponding to given space group.
}
\examples{
# P1 is part of the TRICLINIC family
crystal_family(1)

# The object returned is a string
cfam <- crystal_family(1)
class(cfam)
}
