% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clattice.R
\name{crystal_system}
\alias{crystal_system}
\title{Crystal system corresponding to given space group.}
\usage{
crystal_system(gn)
}
\arguments{
\item{gn}{A natural integer (1,2,3,...). the space group number.}
}
\value{
A character string, the name of the crystal system associated
        with the given space group. If the input integer does not correspond
        any existing space group, the function returns NULL and throws a
        warning.
}
\description{
Crystal system corresponding to given space group.
}
\examples{
# P1 is part of the TRICLINIC system
crystal_system(1)

# The object returned is a string
csys <- crystal_system(1)
class(csys)
}
