% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHELX_dot.R
\name{.exportCSV}
\alias{.exportCSV}
\title{Export shelx table to csv file.}
\usage{
.exportCSV(filenameDF, path, message = TRUE)
}
\arguments{
\item{filenameDF}{A data frame. The name of the sheld data sets to plot CCall vs Ccweak}

\item{path}{Path where the file will be saved.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
plot
}
\description{
Export shelx table to csv file.
}
\examples{
\dontrun{
filename <- shelxd_dataframe
ltmp <- .plotSHELXD(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
