% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XDS_dot.R
\name{.extract_RE_xds}
\alias{.extract_RE_xds}
\title{Reads XDS file file}
\usage{
.extract_RE_xds(filename, message = TRUE)
}
\arguments{
\item{filename}{A character string. The path to a valid shelx log file.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
A named list. Each name correspond to a valid field in the shelx log
   header.
}
\description{
Reads XDS file file
}
\examples{
\dontrun{
filename <- "path/to/a/valid/xds/file.HKL"
ltmp <- .readXDS(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
