% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHELX_dot.R
\name{.plotSHELXC}
\alias{.plotSHELXC}
\title{Plot SHELXC log file}
\usage{
.plotSHELXC(filenameDF, var, message = TRUE)
}
\arguments{
\item{filenameDF}{A character string. The name of the shelc data frame.}

\item{var}{The variable to be plotted against the resolution}

\item{ylabel}{The y label.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
plot
}
\description{
Plot SHELXC log file
}
\examples{
\dontrun{
filename <- shelxc_dataframe
ltmp <- .plotSHELXC(filenameDF)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
