% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHELX_dot.R
\name{.plotSHELXD}
\alias{.plotSHELXD}
\title{Plot SHELXD log file}
\usage{
.plotSHELXD(filenameDF, message = TRUE)
}
\arguments{
\item{filenameDF}{A list containg the SHELXE data frame. The name of the shelxe data sets}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
plot
}
\description{
Plot SHELXD log file
}
\examples{
\dontrun{
filename <- shelxd_dataframe
ltmp <- .plotSHELXD(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
