% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHELX_dot.R
\name{.plotSHELXE}
\alias{.plotSHELXE}
\title{Plot SHELXE log file}
\usage{
.plotSHELXE(filenameDF_i, filenameDF_o, message = TRUE)
}
\arguments{
\item{filenameDF}{A character string. The name of the sheld data sets to plot CCall vs Ccweak}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
plot
}
\description{
Plot SHELXE log file
}
\examples{
\dontrun{
filename <- shelxd_dataframe
plot <- .plotSHELXD(filename)
}
}
