% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strufac.R
\name{expand_au}
\alias{expand_au}
\title{Calculate symmetry-equivalent positions}
\usage{
expand_au(xyzf, aname, B, Occ, SG, inau = FALSE)
}
\arguments{
\item{xyzf}{An n X 3 array or data frame whose rows are the fractional
coordinates of the n atoms in the asymmetric unit.}

\item{aname}{A character string. The name of the atomic species.}

\item{B}{A vector of length n providing the thermal factor for the n
atoms in the asymmetric unit.}

\item{Occ}{A vector of length n providing the occupancy numbers for the
n atoms in the asymmetric unit.}

\item{SG}{A character string indicating the extended
Hermann-Mauguin symbol for the space group.}

\item{inau}{Logical variable. If TRUE, all atoms outside the unit cell
are changed into atoms inside the unit cell by translational
repetition (default is not to change expanded atoms).}
}
\value{
set An integer equal to the specific setting
            corresponding to the given xHM symbol.

An m X 6 data frame with column names x, y, z, atom, B, Occ,
        where m is the total number of atoms after the symmetry-expansion
        and the centring. The first three columns contain the three
        fractional coordinates, while the last three contain the atom name,
        the B factor and the occupancy.
}
\description{
Function to calculate all positions related by symmetry (including cell
centring) to the position of an atom in the asymmetric unit
}
\details{
One atom has fractional coordinates \eqn{(x,y,z)}. The function first
finds the symmetry-equivalent \eqn{(x',y',z')} based on the operation,
\deqn{
  \left(\begin{array}{c}
    x' \\ y' \\ z'
  \end{array}\right) = R
  \left(\begin{array}{c}
    x' \\ y' \\ z'
  \end{array}\right) + T
}
where \eqn{R} is a 3 X 3 matrix representing the point group operation
and \eqn{T} a 3 X 1 vector representing the group translation. The
function then finds the copies \eqn{(x'',y'',z'')} of the symmetry-related
atoms \eqn{(x',y',z')} using the centring operator 3 X 1 vector \eqn{C}:
\deqn{
  \left(\begin{array}{c}
    x'' \\ y'' \\ z''
  \end{array}\right) =
  \left(\begin{array}{c}
    x' \\ y' \\ z'
  \end{array}\right) + C
}
}
\examples{

# Create a "pretend" structure with five atoms in C 1 2 1
# Asymmetric unit is 0<=x<=1/2; 0<=y<1/2; 0<=z<1
xyzf <- matrix(nrow=5,ncol=3)
xyzf[1,] <- c(0.2,0.1,0.6)
xyzf[2,] <- c(0.15,0.15,0.55)
xyzf[3,] <- c(0.05,0.2,0.4)
xyzf[4,] <- c(0.25,0.15,0.56)
xyzf[5,] <- c(0.3,0.1,0.7)

# Atom names
aname <- c("C","O","C","N","N")

# Random B factors
B <- rnorm(5,mean=1,sd=0.2)

# Occupancies
Occ <- rep(1,times=5)

# Space group
SG <- "C 1 2 1"

# Expansion
xyz <- expand_au(xyzf,aname,B,Occ,SG)
print(xyz) # The expanded structure should have 20 atoms

}
