% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections.R
\name{expand_miller}
\alias{expand_miller}
\title{Miller indices expansion by symmetry}
\usage{
expand_miller(SG, hkl)
}
\arguments{
\item{SG}{A character string or a number indicating the
extended Hermann-Mauguin symbol for the space group.}

\item{hkl}{An array or data frame including Miller indices, h, k, l.
These are the reflections to be expanded by symmetry.}
}
\value{
A data frame with the same structure as the
        input object, with the number of rows increased to
        accommodate the reflections generated by symmetry.
        A column of integers is added to the output array,
        indicating a unique number for all symmtry-equivalent
        reflections.
}
\description{
Expand an array including Miller indices according to all
operations of the given symmetry (point) group
}
\details{
The input array contains the three Miller indices, h, k, l.
The array will be expanded to accommodate all equivalent reflections,
according to the point group. Duplicates are removed and no check is
done to detect the presence of systematic absences, which must not be
included in the input array. A column is added to the output array,
indicating the number (1 to nsym) of the symmetry operator producing
the specific set of Miller indices. This is useful to single out all
symmetry-equivalent reflections.
}
\examples{
# Three reflections to be expanded with "P 4"
hkl <- matrix(c(0,0,0,1,0,0,0,1,0),ncol=3)

}
