% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHELX_dot.R
\docType{data}
\name{filename}
\alias{filename}
\title{Reads SHELX log file file}
\format{An object of class \code{character} of length 1.}
\usage{
filename
}
\arguments{
\item{filename}{A character string. The path to a valid shelx log file.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the shelx log header.}
}
\value{
A named list. Each name correspond to a valid field in the shelx log
   header.
}
\description{
Reads SHELX log file file
}
\examples{
\dontrun{
filename <- "path/to/a/valid/shelxlog/file.log"
ltmp <- .readXDS(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
\keyword{datasets}

