% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dot.R
\name{\%X\%}
\alias{\%X\%}
\title{Cross product u X v in 3D. Returns a 3 components vector}
\usage{
u \%X\% v
}
\arguments{
\item{u}{A real numeric vector of length 3.}

\item{v}{A real numeric vector of length 3.}
}
\value{
A real numeric vector of length 3. This is the cross product of
         vectors u and v.
}
\description{
Cross product u X v in 3D. Returns a 3 components vector
}
\examples{
u <- c(1,0,0)
v <- c(0,1,0)
w <- u\%X\%v
print(w)
}
