% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{invfousynth3D}
\alias{invfousynth3D}
\title{Inverse Fourier synthesis}
\usage{
invfousynth3D(rho, hkl)
}
\description{
From density to structure factors
}
\details{
The electron density in the unit cell, \eqn{\rho(x,y,z)}, where
\eqn{(x,y,z)} are fractional coordinates, can be computed using
amplitudes and phases of the structure factors. The synthesis
formula is:
 \deqn{
  \rho(x,y,z) = \frac{1}{V}\sum_{h,k,l} F(h,k,l)
  \exp(-2\pi\mathrm{i}(hx+ky+lz))
}
where \eqn{F(h,k,l)} is the structure factor for the set of Miller indices
\eqn{h,k,l} and \eqn{V} is the unit cell volume. The inverse Fourier
synthesis yields the structure factors from the electron density.
}
