% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections_S3.R
\name{merged_reflections}
\alias{merged_reflections}
\title{Constructor for an S3 object of class "merged_reflections".}
\usage{
merged_reflections(ruc = NULL, csym = NULL, records = NULL,
  dtypes = NULL, set = 1)
}
\arguments{
\item{ruc}{An object of class "rec_unit_cell" (which represents
a reciprocal unit cell).}

\item{csym}{An object of class "cryst_symm" (which represents
a crystallographic symmetry group).}

\item{records}{A data frame containing all reflections coming
from the x-ray data collection on the crystal. This
data frame must include at least the three Miller
indices, H, K, L (of dtype "H").}

\item{dtypes}{A character vector whose length is the same as
the number of columns in 'records'. One character (a
capital letter) is associated with each type of data.
For example, a Miller index is of dtype "H"; a structure
amplitude is of dtype "F"; an anomalous difference is of
dtype "D"; etc (see details later).}

\item{set}{An integer number corresponding to the specific
setting for the given space group. Default is 1.}
}
\value{
An object of class "merged_reflections". It is a named
        list of length 4 whose names are:
        \describe{
          \item{ruc}{An object of class "rec_unit_cell".}
          \item{csym}{An object of class "cryst_symm".}
          \item{records}{A data frame containing the data.}
          \item{dtypes}{A character vector containing the
                type of data (Miller indices, structure
                factors, etc).}
        }
}
\description{
This represents scaled and merged x-ray data from one crystal.
}
\details{
If the constructor is used without arguments, the default
object created will be composed of 3 reflections, (1,0,0),
(0,1,0), (0,0,1) from a cubic crystal with cell of side 10
angstroms, and symmetry P 2 3. The only available columns
will be of dtype "H", and named H, K, L (the Miller indices).
}
\examples{
# Create an orthorombic (default) cell
uc <- unit_cell(10,30,15)

# Create the related reciprocal cell
ruc <- create_rec_unt_cell(uc)

# Create symmetry (P n c 2)
csym <- cryst_symm(30)

# Create a few records
records <- expand.grid(H=-2:2,K=-2:2,L=-2:2)

# Create merged_reflections object with H, K, L
mrefs <- merged_reflections(ruc,csym,records,set=1)

}
