% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSHELXC.R
\name{plotSHELXC}
\alias{plotSHELXC}
\title{Plot SHELXC log files}
\usage{
plotSHELXC(filenameDF, var, message = TRUE)
}
\arguments{
\item{filenameDF}{A data frame containing the variable from shelxc.}

\item{var}{the variable to be plotted vs the resolution}
}
\value{
A graphics containg the solution founded by SHELXD.
}
\description{
Plot SHELXC log files
}
\examples{
\dontrun{
plot <- plotSHELXC(filenameDF, var, message = TRUE)
plot
}
}
