% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSHELXD.R
\name{plotSHELXD}
\alias{plotSHELXD}
\title{Plot SHELXD log files}
\usage{
plotSHELXD(filenameDF)
}
\arguments{
\item{filenameDF}{A data frame containing CCall, CCweak}
}
\value{
A graphics containg the solution founded by SHELXD.
}
\description{
Plot SHELXD log files
}
\examples{
\dontrun{
plot <- plotSHELXD(filenameDF)
plot
}
}

