% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSHELXE.R
\name{plotSHELXE}
\alias{plotSHELXE}
\title{Plot SHELXD log files}
\usage{
plotSHELXE(filenameDF_i, filenameDF_o, message = TRUE)
}
\arguments{
\item{filenameDF_i}{A data frame containing the inverted hand.}

\item{filenameDF_o}{A data frame containing the original hand.}
}
\value{
A graphics containg the solution founded by SHELXD.
}
\description{
Plot SHELXD log files
}
\examples{
\dontrun{
plot <- plotSHELXE(filenameDF_i, filenameDF_o, message = TRUE)
plot
}
}

