% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_S3.R
\name{print.angle}
\alias{print.angle}
\title{Print method for an object of class "angle".}
\usage{
\method{print}{angle}(x, ...)
}
\arguments{
\item{x}{An object of class "angle".}

\item{...}{Additional arguments passed to the print methods}
}
\description{
The value is displayed in degrees
}
\examples{
# Create an angle of 90 degrees using radians
ang1 <- angle(pi/2,FALSE)

# Display its value
print(ang1)
}
