% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_S3.R
\name{print.bravais}
\alias{print.bravais}
\title{Print method for an object of class "bravais".}
\usage{
\method{print}{bravais}(x, ...)
}
\arguments{
\item{x}{An object of class "bravais".}

\item{...}{Additional arguments passed to the print methods}
}
\description{
The Bravais lattice and related crystal family, crystal system and lattice system
are displayed.
}
\examples{
# Create a triclinic Bravais lattice
bt <- bravais()

# Display its value
print(bt)
}
