% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_S3.R
\name{print.CIF}
\alias{print.CIF}
\title{Print method for an object of class "CIF".}
\usage{
\method{print}{CIF}(x, ...)
}
\arguments{
\item{x}{An object of class "CIF".}

\item{...}{Additional arguments passed to the print methods}
}
\description{
CIF files include data related to crystallography in a popular framework supported by
IUCr, known as Crystallographic Information Framework. The content included varies according
to the type of data stored and it is organised in a list of objects, each one with a given
standard name. A few, commonly-used names are HEADER, SYMM, REFL, COOR, etc.
}
\examples{
# Create a CIF object from a data file contained in the cry package
# cobj <- CIF("directory/file.cif")

# Display its value
#print(cobj)
}
