% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_S3.R
\name{print.cryst_symm}
\alias{print.cryst_symm}
\title{Print method for an object of class "cryst_symm".}
\usage{
\method{print}{cryst_symm}(x, ...)
}
\arguments{
\item{x}{An object of class "cryst_symm".}

\item{...}{Additional arguments passed to the print methods}
}
\description{
xxx
}
\examples{
# Create an object of P 2 symmetry
crsym <- cryst_symm("P 2")

# Display its value
print(crsym)

}
