% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R, R/reflections_S3.R
\name{print.unit_cell}
\alias{print.unit_cell}
\alias{print.merged_reflections}
\title{Print method for an object of class "unit_cell".}
\usage{
\method{print}{unit_cell}(x, ...)

\method{print}{merged_reflections}(x, ...)
}
\arguments{
\item{x}{An object of class "merged_reflections".}

\item{...}{Additional arguments passed to the print methods}
}
\description{
The values are displayed in angstroms and degrees

The output includes details on the unit cell, the crystal
symmetry and the first 10 records (data).
}
\examples{
# Create a cubic unit cell
uc <- unit_cell(10,10,10,90,90,90)

# Display its value
print(uc)
# Create a default 'merged_reflections' object
mrefs <- merged_reflections()

# Display its value
print(mrefs)

}
