% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SHELX_log.R
\name{read_SHELX_log}
\alias{read_SHELX_log}
\title{Reads and SHELXD log files}
\usage{
read_SHELX_log(filename)
}
\arguments{
\item{filename}{A character string. The path to a valid log file.}
}
\value{
A named list. Each name correspond to a valid field in the log
   header.
}
\description{
Reads and SHELXD log files
}
\examples{
datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"shelxc.log")
ltmp <- read_SHELX_log(filename)
print(names(ltmp))

}
