% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif.R
\name{readCIF}
\alias{readCIF}
\title{Reads and output an CIF file}
\usage{
readCIF(filename, messages = FALSE)
}
\arguments{
\item{filename}{A character string. The path to a valid CIF file.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the cif file.}
}
\value{
A named list. Each name correspond to a valid field in the cif.
}
\description{
Reads and output an CIF file
}
\examples{
#filename <- "/path_to_file/example.cif"
#lcif <- readCIF(filename)
#print(names(lcif))
#print(lcif$INTRO$CELL)
#print(lcif$INTRO$HALL)
#print(lcif$INTRO$HM)
#print(lcif$SYMM)
}
