% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCORRECT.R
\name{readCORRECT}
\alias{readCORRECT}
\title{Reads XDS CORRECT file, check if there are alien
reflections and create a REMOVE.HKL.}
\usage{
readCORRECT(filename, message = TRUE, alien = TRUE, path)
}
\arguments{
\item{filename}{A character string. The path to a valid xds file.}

\item{alien}{A logical variable, which the default value is TRUE. This variable}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the xds header.}
}
\value{
A named list. Each name correspond to a valid field in the xds
   header.
}
\description{
ISIGMA will cut the data following the wished I over sigma
}
\examples{
\dontrun{
filename <- "path/to/a/valid/HKL/CORRECT.LP"
correct <- readCORRECT(filename, alien = TRUE)
}
}
