% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSHELXlog.R
\name{readSHELXlog}
\alias{readSHELXlog}
\title{Reads and SHELXD log files}
\usage{
readSHELXlog(filename)
}
\arguments{
\item{filename}{A character string. The path to a valid log file.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the log header.}
}
\value{
A named list. Each name correspond to a valid field in the log
   header.
}
\description{
Reads and SHELXD log files
}
\examples{
\dontrun{
filename <- "path/to/a/valid/shelxd/shelxd.log"
ltmp <- .readSHELXDlog(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
}
