% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXSCALE.R
\name{readXSCALE}
\alias{readXSCALE}
\title{Reads and output XSCALE.ahkl file with header and reflections}
\usage{
readXSCALE(filename, messages = TRUE, ISIGMA = 1)
}
\arguments{
\item{filename}{A character string. The path to a valid mtz file.}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the mtz header.}
}
\value{
A named list. Each name correspond to a valid field in the mtz
   header.
}
\description{
Reads and output XSCALE.ahkl file with header and reflections
}
\examples{
\dontrun{
filename <- "path/to/a/valid/ahkl/file.ahkl"
ltmp <- readXSCALE(filename)
print(names(ltmp))
print(ltmp$UNIT_CELL_CONSTANTS)
}
}

