% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{rec_unit_cell}
\alias{rec_unit_cell}
\title{Constructor for an S3 object of class "rec_unit_cell".}
\usage{
rec_unit_cell(ar, br, cr, aar, bbr, ccr)
}
\arguments{
\item{ar}{A real number. One of the reciprocal unit cell's side lengths, in 1/angstroms.}

\item{br}{A real number. One of the reciprocal unit cell's side lengths, in 1/angstroms.}

\item{cr}{A real number. One of the reciprocal unit cell's side lengths, in 1/angstroms.}

\item{aar}{A real number. One of the reciprocal unit cell's angles, in degrees.}

\item{bbr}{A real number. One of the reciprocal unit cell's angles, in degrees.}

\item{ccr}{A real number. One of the reciprocal unit cell's angles, in degrees.}
}
\value{
An object of class "rec_unit_cell". It is a named list of length 6 whose
        last three slots are of "angle" class.
}
\description{
This represents a crystal reciprocal unit cell.
}
\examples{
# Create a cubic reciprocal unit cell
ruc1 <- rec_unit_cell(1/10,1/10,1/10,90,90,90)
class(ruc1)
}
