% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections.R
\name{reduce_miller}
\alias{reduce_miller}
\title{Unique set of Miller indices}
\usage{
reduce_miller(SG, hkl)
}
\arguments{
\item{SG}{A character string or a number indicating the
extended Hermann-Mauguin symbol for the space group.}

\item{hkl}{An array or data frame including Miller indices h,k,l
These are the reflections to be reduced.}
}
\value{
An data frame with the same structure as the
        input object, with the number of rows decreased to
        include only the unique reflections for the specific
        space group. A fourth column indicating the number of
        symmetry operator is added, if initially missing.
}
\description{
Reduces an array including Miller indices to the asymmetric
unit in reciprocal space (unique reflections).
}
\details{
The reflections array will be reduced to the set of unique reflections,
depending on the space group of the structure. No checks are done to
detect the presence of systematic absences, which must not be included
in the input array.
}
\examples{
# Expand two unique reflections with "P 4"
hkl <- data.frame(H=c(2,1),K=c(2,1),L=c(1,0))
print(hkl)
SG <- "P 4"
new_hkl <- expand_miller(SG,hkl)
print(new_hkl)

# Now reduce the new array: it shoul dreturn the original one
hkl2 <- reduce_miller(SG,new_hkl)
print(hkl2)

}
