% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strufac.R
\name{scafac}
\alias{scafac}
\title{Atomic scattering factor}
\usage{
scafac(atom_name, s)
}
\arguments{
\item{atom_name}{A character string. The symbol of the atom whose
scattering factor needs to be calculated (see list of
available names in \code{\link{sfcoeffs}}).}

\item{s}{A real numeric vector, the values of \eqn{s} at which
the scattering function needs to be evaluated.}
}
\value{
The value of the scattering factor at s.
}
\description{
The atomic scattering factor function used in x-ray
crystallography
}
\details{
The scattering factor is normally built as a function of
\eqn{\sigma\equiv\sin\theta/\lambda=s/2}, where \eqn{s}
is the length of the Miller indices (as a vector) in reciprocal
space, \eqn{s=|\mathbf{h}|}. But in the expression for the
structure factor it appears as a function of \eqn{s}, rather
than a function of \eqn{\sigma}. Here the scattering function
is a function of \eqn{s}. The approximation used here (and in
most crystallography packages) is accurate for up to 2 angstroms
resolution. It can be used for resolutions higher than 2
angstroms, but it is less accurate.
}
\examples{

# Resolution range (up to 2 angstroms)
s <- seq(0,0.5,length.out=1000)

# The chosen atom is nitrogen
atom_name <- "N"

# Scattering factor
fN <- scafac(atom_name,s)

# Plot

}
\seealso{
\code{\link{sfcoeffs}}
}
