% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strufac.R
\name{sfcoeffs}
\alias{sfcoeffs}
\title{Coefficients for scattering factors}
\usage{
sfcoeffs(atom_name)
}
\arguments{
\item{atom_name}{A character string. The symbol of the atom whose
scattering factor's coefficients are needed (see list of
available names in details).}
}
\value{
A named list whose three components are vectors and
        scalars containing the coefficients:
        \describe{
          \item{a}{Vector of length 4 containing the \eqn{a_j}}
          \item{b}{Vector of length 4 containing the \eqn{b_j}}
          \item{c}{Single number, the value of \eqn{c}}
        }
}
\description{
The nine coefficients of the four-gaussians approximation for
the scattering factors
}
\details{
A scattering factor is a function of
\eqn{\sigma\equiv\sin(\theta)/\lambda},
which is equal to \eqn{s/2}, with \eqn{s=1/d} and \eqn{d} the
resolution of the given reflection. The scattering form is
satisfactorily modelled as sum of four gaussians:
\deqn{
a_1\exp(-b_1 \sigma^2) + a_2\exp(-b_2 \sigma^2) +
a_3\exp(-b_3 \sigma^2) + a_4\exp(-b_4 \sigma^2) + c
}
This function returns the coefficients as a named list with
three components, \eqn{a=(a_1,a_2,a_3,a_4)},
\eqn{b=(b_1,b_2,b_3,b_4)} and \eqn{c=c}. The possible atom names
(also included are some ionic species) are here summarized.
\itemize{
\item H
\item H-1
\item He
\item Li
\item Li+1
\item Be
\item Be+2
\item B
\item C
\item Cv
\item N
\item O
\item O-1
\item F
\item F-1
\item Ne
\item Na
\item Na+1
\item Mg
\item Mg+2
\item Al
\item Al+3
\item Si
\item Siv
\item Si+4
\item P
\item S
\item Cl
\item Cl-1
\item Ar
\item K
\item K+1
\item Ca
\item Ca+2
\item Sc
\item Sc+3
\item Ti
\item Ti+2
\item Ti+3
\item Ti+4
\item V
\item V+2
\item V+3
\item V+5
\item Cr
\item Cr+2
\item Cr+3
\item Mn
\item Mn+2
\item Mn+3
\item Mn+4
\item Fe
\item Fe+2
\item Fe+3
\item Co
\item Co+2
\item Co+3
\item Ni
\item Ni+2
\item Ni+3
\item Cu
\item Cu+1
\item Cu+2
\item Zn
\item Zn+2
\item Ga
\item Ga+3
\item Ge
\item Ge+4
\item As
\item Se
\item Br
\item Br-1
\item Kr
\item Rb
\item Rb+1
\item Sr
\item Sr+2
\item Y
\item Y+3
\item Zr
\item Zr+4
\item Nb
\item Nb+3
\item Nb+5
\item Mo
\item Mo+3
\item Mo+5
\item Mo+6
\item Tc
\item Ru
\item Ru+3
\item Ru+4
\item Rh
\item Rh+3
\item Rh+4
\item Pd
\item Pd+2
\item Pd+4
\item Ag
\item Ag+1
\item Ag+2
\item Cd
\item Cd+2
\item In
\item In+3
\item Sn
\item Sn+2
\item Sn+4
\item Sb
\item Sb+3
\item Sb+5
\item Te
\item I
\item I-1
\item Xe
\item Cs
\item Cs+1
\item Ba
\item Ba+2
\item La
\item La+3
\item Ce
\item Ce+3
\item Ce+4
\item Pr
\item Pr+3
\item Pr+4
\item Nd
\item Nd+3
\item Pm
\item Pm+3
\item Sm
\item Sm+3
\item Eu
\item Eu+2
\item Eu+3
\item Gd
\item Gd+3
\item Tb
\item Tb+3
\item Dy
\item Dy+3
\item Ho
\item Ho+3
\item Er
\item Er+3
\item Tm
\item Tm+3
\item Yb
\item Yb+2
\item yb+3
\item Lu
\item Lu+3
\item Hf
\item Hf+4
\item Ta
\item Ta+5
\item W
\item W+6
\item Re
\item Os
\item Os+4
\item Ir
\item Ir+3
\item Ir+4
\item Pt
\item Pt+2
\item Pt+4
\item Au
\item Au+1
\item Au+3
\item Hg
\item Hg+1
\item hg+2
\item Tl
\item Tl+1
\item Tl+3
\item Pb
\item Pb+2
\item Pb+4
\item Bi
\item Bi+3
\item Bi+5
\item Po
\item At
\item Rn
\item Fr
\item Ra
\item Ra+2
\item Ac
\item Ac+3
\item Th
\item Th+4
\item Pa
\item U
\item U+3
\item U+4
\item U+6
\item Np
\item Np+3
\item Np+4
\item Np+6
\item Pu
\item Pu+3
\item Pu+4
\item Pu+6
\item Am
\item Cm
\item Bk
\item Cf
}
}
\examples{

# Coefficients for the carbon atom
aname <- "C"
lcoeffs <- sfcoeffs(aname)
print(lcoeffs)

}
