% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cry.R
\name{theme_cry}
\alias{theme_cry}
\title{cry theme for ggplot2}
\usage{
theme_cry()
}
\description{
cry theme for ggplot2
}
\examples{
\dontrun{
 plot_SHELX(obj_shelxc, var = obj_shelxc$Chi_sq, type = "shelxc",
 title_chart = "Chis ^2") + theme_cry
}
}
