% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cry_generics_S3.R
\name{create_unit_cell}
\alias{create_unit_cell}
\title{S3 generic to create unit_cell objects}
\usage{
create_unit_cell(a, ...)
}
\arguments{
\item{a}{An object or objects used to select a method. These can be
cell parameters, an object of class rec_unit_cell, etc.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class "unit_cell". It is a named list of length 6 whose
        last three slots are of class "angle".
}
\description{
The unit_cell object can be created starting from specific objects, files, etc.
}
\examples{
# Create a unit_cell in default (no arguments)
uc <- create_unit_cell()
print(uc)
}
