% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_S3.R
\name{print.unit_cell}
\alias{print.unit_cell}
\title{Print method for an object of class "unit_cell".}
\usage{
\method{print}{unit_cell}(x)
}
\arguments{
\item{x}{An object of class "unit_cell"}
}
\description{
The values are displayed in angstroms and degrees
}
\examples{
# Create a cubic unit cell
uc <- unit_cell(10,10,10,90,90,,90)

# Display its value
print(uc)
}
