% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections.R
\name{readMTZHeader}
\alias{readMTZHeader}
\title{Reads and output an MTZ header}
\usage{
readMTZHeader(filename, message = FALSE)
}
\arguments{
\item{filename}{A character string. The path to a valid mtz file.}

\item{message}{A logical variable. If TRUE the function prints
a message highlighting what is included in the mtz header.
Default value is \code{message=FALSE}.}
}
\value{
A named list. Each name correspond to a valid field in the mtz
   header.
}
\description{
Reads and output an MTZ header
}
\examples{
datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"insulin_merged.mtz")
ltmp <- readMTZHeader(filename)
print(names(ltmp))
print(ltmp$CELL)
print(ltmp$SYMM)
}
