% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSHELXlog.R
\name{readSHELXlog}
\alias{readSHELXlog}
\title{Reads and SHELXD log files}
\usage{
readSHELXlog(filename)
}
\arguments{
\item{filename}{A character string. The path to a valid log file.}
}
\value{
A named list. Each name correspond to a valid field in the log
   header.
}
\description{
Reads and SHELXD log files
}
\examples{
datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"shelxc.log")
ltmp <- readSHELXlog(filename)
print(names(ltmp))

}
