% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXDS_ASCII.R
\name{readXDS_ASCII}
\alias{readXDS_ASCII}
\title{Reads XDS file with header and reflections.}
\usage{
readXDS_ASCII(filename, message = TRUE, ISIGMA = 1)
}
\arguments{
\item{filename}{A character string. The path to a valid xds file.}

\item{ISIGMA}{A numerical value, which the default value is 1}

\item{messages}{A logical variable. If TRUE (default) the function prints
a message highlighting what is included in the xds header.}
}
\value{
A named list. Each name correspond to a valid field in the xds
   header.
}
\description{
ISIGMA will cut the data following the wished I over sigma
}
\examples{
\dontrun{
filename <- "path/to/a/valid/HKL/file.HKL"
xds_ascii <- readXDS(filename)
}
}
